requirements.txt
# Criminology Management System - Python Dependencies
# Install with: pip install -r requirements.txt

# Core Flask Framework
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
WTForms==3.1.2

# Database Drivers
PyMySQL==1.1.0
mysqlclient==2.2.4

# Security
Werkzeug==3.1.3
itsdangerous==2.2.0

# Template Engine
Jinja2==3.1.6
MarkupSafe==3.0.2

# Utilities
click==8.2.1
blinker==1.9.0

# Development Dependencies (optional)
# Uncomment if needed for development
# python-dotenv==1.0.0
# flask-migrate==4.0.5
# flask-cors==4.0.0